add SQLBackend interface
