+++ cygnus-ngsi/src/main/java/com/telefonica/iot/cygnus/sinks/NGSIPostgisSink.java
@@ -21,7 +21,7 @@ package com.telefonica.iot.cygnus.sinks;
 import com.telefonica.iot.cygnus.aggregation.NGSIGenericAggregator;
 import com.telefonica.iot.cygnus.aggregation.NGSIGenericColumnAggregator;
 import com.telefonica.iot.cygnus.aggregation.NGSIGenericRowAggregator;
-import com.telefonica.iot.cygnus.backends.sql.SQLBackend;
+import com.telefonica.iot.cygnus.backends.sql.SQLBackendImpl;
 import com.telefonica.iot.cygnus.errors.CygnusBadConfiguration;
 import com.telefonica.iot.cygnus.errors.CygnusBadContextData;
 import com.telefonica.iot.cygnus.errors.CygnusCappingError;
@@ -67,7 +67,7 @@ public class NGSIPostgisSink extends NGSISink {
     private String postgisPassword;
     private boolean rowAttrPersistence;
     private int maxPoolSize;
-    private static volatile SQLBackend postgisPersistenceBackend;
+    private static volatile SQLBackendImpl postgisPersistenceBackend;
     private boolean enableCache;
     private boolean swapCoordinates;
     private boolean attrNativeTypes;
@@ -150,7 +150,7 @@ public class NGSIPostgisSink extends NGSISink {
      * Returns the persistence backend. It is protected due to it is only required for testing purposes.
      * @return The persistence backend
      */
-    protected SQLBackend getPersistenceBackend() {
+    protected SQLBackendImpl getPersistenceBackend() {
         return postgisPersistenceBackend;
     } // getPersistenceBackend
 
@@ -158,7 +158,7 @@ public class NGSIPostgisSink extends NGSISink {
      * Sets the persistence backend. It is protected due to it is only required for testing purposes.
      * @param postgisPersistenceBackend
      */
-    protected void setPersistenceBackend(SQLBackend postgisPersistenceBackend) {
+    protected void setPersistenceBackend(SQLBackendImpl postgisPersistenceBackend) {
         this.postgisPersistenceBackend = postgisPersistenceBackend;
     } // setPersistenceBackend
 
@@ -269,7 +269,7 @@ public class NGSIPostgisSink extends NGSISink {
      */
     private static synchronized void createPersistenceBackend(String sqlHost, String sqlPort, String sqlUsername, String sqlPassword, int maxPoolSize, String defaultSQLDataBase) {
         if (postgisPersistenceBackend == null) {
-            postgisPersistenceBackend = new SQLBackend(sqlHost, sqlPort, sqlUsername, sqlPassword, maxPoolSize, POSTGIS_INSTANCE_NAME, POSTGIS_DRIVER_NAME, defaultSQLDataBase);
+            postgisPersistenceBackend = new SQLBackendImpl(sqlHost, sqlPort, sqlUsername, sqlPassword, maxPoolSize, POSTGIS_INSTANCE_NAME, POSTGIS_DRIVER_NAME, defaultSQLDataBase);
         }
     }
